[Fixed] - Some warnings

+++ Game/src/World/World.cpp
@@ -77,8 +77,8 @@ void World::Update()
 		{
 			glm::vec2 chunkPos = glm::vec2(it->first.m_X, it->first.m_Y);
 			glm::vec2 camPos = glm::vec2(-xCamGridPos, -zCamGridPos);
-			int length = glm::length(chunkPos - camPos);
-			if (static_cast<float>(length) > m_LoadDistance * 0.5f)
+			float length = glm::length(chunkPos - camPos);
+			if (length > m_LoadDistance * 0.5f)
 			{
 				it->second.Unload();
 				m_ChunkMap.erase(it++);
